Work with editor plugin.
