+++ Config/DefaultEditorPerProjectUserSettings.ini
@@ -1,2 +1,6 @@
 [ContentBrowser]
 ContentBrowserTab1.SelectedPaths=/Game/FirstPerson
+
+[/Script/SourceCodeAccess.SourceCodeAccessSettings]
+PreferredAccessor=VisualStudio2017
+
