--- Plugins/ShaderCopyHelper/Source/Public/ShaderFileVisitor.h
@@ -1,18 +1,0 @@
-#pragma once
-
-#include "Engine.h"
-#include "GenericPlatformFile.h"
-
-class FShaderFileVisitor : public IPlatformFile::FDirectoryVisitor
-{
-public:
-	FShaderFileVisitor();
-
-public:
-	virtual bool Visit(const TCHAR* FilenameOrDirectory, bool bIsDirectory) override;
-
-public:
-	void Reset();
-
-	TArray<FString> ShaderFilePaths;
-};
