+++ Plugins/ShaderEngine/Source/ShaderEngine/Public/ShaderEngine.h
@@ -12,7 +12,7 @@
 #include "PixelShader/PixelShaderDeclaration.h"
 #include "PixelShader/PixelShaderUsageExample.h"
 
-class FShaderEngineModule : public IModuleInterface
+class SHADERENGINE_API FShaderEngineModule : public IModuleInterface
 {
 public:
 	virtual void StartupModule() override;
